+++ examples/xlnet/.gitignore
@@ -1,3 +1,2 @@
 /data/
-/xlnet_pretrained_models/
 /processed_data/
