+++ examples/xlnet/utils/processor.py
@@ -13,11 +13,13 @@
 # limitations under the License.
 
 """
-Data processor base class. Adapted from
-https://github.com/zihangdai/xlnet/blob/master/run_classifier.py#L144-L192
+Data processors. Adapted from
+https://github.com/zihangdai/xlnet/blob/master/run_classifier.py
 """
 
 import csv
+import logging
+from abc import ABC
 from pathlib import Path
 from typing import NamedTuple, Optional, Union, List, Dict, Type
 
@@ -96,3 +98,198 @@ def get_processor_class(task: str) -> Type[DataProcessor]:
     if klass is None:
         raise ValueError(f"Unsupported task {task}")
     return klass
+
+
+class GLUEProcessor(DataProcessor, ABC):
+    train_file = "train.tsv"
+    dev_file = "dev.tsv"
+    test_file = "test.tsv"
+    label_column: int
+    text_a_column: int
+    text_b_column: int
+    contains_header = True
+    test_text_a_column: int
+    test_text_b_column: int
+    test_contains_header = True
+
+    def __init__(self, data_dir: str):
+        super().__init__(data_dir)
+        if not hasattr(self, 'test_text_a_column'):
+            self.test_text_a_column = self.text_a_column
+        if not hasattr(self, 'test_text_b_column'):
+            self.test_text_b_column = self.text_b_column
+
+    def get_train_examples(self) -> List[InputExample]:
+        return self._create_examples(
+            self._read_tsv(self.data_dir / self.train_file), "train")
+
+    def get_dev_examples(self) -> List[InputExample]:
+        return self._create_examples(
+            self._read_tsv(self.data_dir / self.dev_file), "dev")
+
+    def get_test_examples(self) -> List[InputExample]:
+        return self._create_examples(
+            self._read_tsv(self.data_dir / self.test_file), "test")
+
+    def _create_examples(self, lines: List[List[str]],
+                         set_type: str) -> List[InputExample]:
+        """Creates examples for the training and dev sets."""
+        examples = []
+        for (i, line) in enumerate(lines):
+            if i == 0 and self.contains_header and set_type != "test":
+                continue
+            if i == 0 and self.test_contains_header and set_type == "test":
+                continue
+            guid = f"{set_type}-{i}"
+
+            a_column = (self.text_a_column if set_type != "test" else
+                        self.test_text_a_column)
+            b_column = (self.text_b_column if set_type != "test" else
+                        self.test_text_b_column)
+
+            # there are some incomplete lines in QNLI
+            if len(line) <= a_column:
+                logging.warning('Incomplete line, ignored.')
+                continue
+            text_a = line[a_column]
+
+            if b_column is not None:
+                if len(line) <= b_column:
+                    logging.warning('Incomplete line, ignored.')
+                    continue
+                text_b = line[b_column]
+            else:
+                text_b = None
+
+            if set_type == "test":
+                label = self.labels[0]
+            else:
+                if len(line) <= self.label_column:
+                    logging.warning('Incomplete line, ignored.')
+                    continue
+                label = line[self.label_column]
+            examples.append(InputExample(guid, text_a, text_b, label))
+        return examples
+
+
+@DataProcessor.register("MNLI", "MNLI_matched")
+class MnliMatchedProcessor(GLUEProcessor):
+    labels = ["contradiction", "entailment", "neutral"]
+
+    dev_file = "dev_matched.tsv"
+    test_file = "test_matched.tsv"
+    label_column = -1
+    text_a_column = 8
+    text_b_column = 9
+
+
+@DataProcessor.register("MNLI_mismatched")
+class MnliMismatchedProcessor(MnliMatchedProcessor):
+    dev_file = "dev_mismatched.tsv"
+    test_file = "test_mismatched.tsv"
+
+
+@DataProcessor.register("STS-B", "stsb")
+class StsbProcessor(GLUEProcessor):
+    labels: List[str] = []
+    is_regression = True
+
+    label_column = 9
+    text_a_column = 7
+    text_b_column = 8
+
+    def _create_examples(self, lines: List[List[str]],
+                         set_type: str) -> List[InputExample]:
+        """Creates examples for the training and dev sets."""
+        examples = []
+        for (i, line) in enumerate(lines):
+            if i == 0 and self.contains_header and set_type != "test":
+                continue
+            if i == 0 and self.test_contains_header and set_type == "test":
+                continue
+            guid = f"{set_type}-{i}"
+
+            a_column = (self.text_a_column if set_type != "test" else
+                        self.test_text_a_column)
+            b_column = (self.text_b_column if set_type != "test" else
+                        self.test_text_b_column)
+
+            # there are some incomplete lines in QNLI
+            if len(line) <= a_column:
+                logging.warning('Incomplete line, ignored.')
+                continue
+            text_a = line[a_column]
+
+            if b_column is not None:
+                if len(line) <= b_column:
+                    logging.warning('Incomplete line, ignored.')
+                    continue
+                text_b = line[b_column]
+            else:
+                text_b = None
+
+            if set_type == "test":
+                label = 0.0
+            else:
+                if len(line) <= self.label_column:
+                    logging.warning('Incomplete line, ignored.')
+                    continue
+                label = float(line[self.label_column])
+            examples.append(InputExample(guid, text_a, text_b, label))
+
+        return examples
+
+
+@DataProcessor.register("Yelp5")
+class Yelp5Processor(DataProcessor):
+    labels = ["1", "2", "3", "4", "5"]
+
+    def get_train_examples(self) -> List[InputExample]:
+        return self._create_examples(self.data_dir / "train.csv")
+
+    def get_dev_examples(self) -> List[InputExample]:
+        return self._create_examples(self.data_dir / "test.csv")
+
+    def get_test_examples(self):  # pylint: disable=no-self-use
+        raise TypeError("The Yelp 5 dataset does not have a test set.")
+
+    @staticmethod
+    def _create_examples(input_file: Path) -> List[InputExample]:
+        """Creates examples for the training and dev sets."""
+        examples = []
+        with input_file.open() as f:
+            reader = csv.reader(f)
+            for i, line in enumerate(reader):
+                label = line[0]
+                text_a = line[1].replace('""', '"').replace('\\"', '"')
+                examples.append(InputExample(
+                    guid=str(i), text_a=text_a, text_b=None, label=label))
+        return examples
+
+
+@DataProcessor.register("IMDB")
+class ImdbProcessor(DataProcessor):
+    labels = ["neg", "pos"]
+
+    def get_train_examples(self) -> List[InputExample]:
+        return self._create_examples(self.data_dir / "train")
+
+    def get_dev_examples(self) -> List[InputExample]:
+        return self._create_examples(self.data_dir / "test")
+
+    def get_test_examples(self):  # pylint: disable=no-self-use
+        raise TypeError("The IMDB dataset does not have a test set.")
+
+    @staticmethod
+    def _create_examples(data_dir: Path) -> List[InputExample]:
+        examples = []
+        for label in ["neg", "pos"]:
+            cur_dir = data_dir / label
+            for filename in cur_dir.iterdir():
+                if filename.suffix != "txt":
+                    continue
+                with (cur_dir / filename).open() as f:
+                    text = f.read().strip().replace("<br />", " ")
+                examples.append(InputExample(
+                    guid=str(filename), text_a=text, text_b=None, label=label))
+        return examples
