+++ examples/xlnet/xlnet_generation_ipython.py
@@ -30,14 +30,12 @@ def main():
         device = 'cpu'
 
     pretrained_model_dir = tx.modules.load_pretrained_xlnet(
-        pretrained_model_name='xlnet-large-cased',
-        cache_dir='xlnet_pretrained_models')
+        pretrained_model_name='xlnet-large-cased')
 
     spm_model_path = os.path.join(pretrained_model_dir, "spiece.model")
 
     model = tx.modules.XLNetDecoder(
-        pretrained_model_name='xlnet-large-cased',
-        cache_dir='xlnet_pretrained_models')
+        pretrained_model_name='xlnet-large-cased')
     model = model.to(device)
 
     sp_model = spm.SentencePieceProcessor()
